List nodes under a meta node

GET /meta/nodes/{meta_node}/list

meta_nodes_list_node(
  meta_node,
  full_data = TRUE,
  limit = 10,
  offset = 0,
  mode = c("table", "raw")
)

Arguments

meta_node
Name of a meta node (e.g. Gwas). Use meta_nodes_list to get the full list of meta nodes.

full_data
When False, only return the id and name fields (their specific names differ in specific nodes) for a node. This is useful if you want your queries to return results faster with smaller amount of data requested.

limit
Max number of items to retrieve.

offset
Number of items to skip. Use limit and offset in combination to do pagination.

mode
If mode = "table", returns a data frame (a tibble as per tidyverse convention). If mode = "raw", returns a raw response from EpiGraphDB API with minimal parsing done by httr.

Value

Data from GET /meta/nodes/{meta_node}/list

Examples

# List the first 5 Gwas nodes, with only id and name fields
meta_nodes_list_node(meta_node = "Gwas", full_data = FALSE, limit = 5)

#> # A tibble: 5 x 2
#>   id                     name                  
#>   <chr>                  <chr>                 
#> 1 eqtl-a-ENSG00000261662 ENSG00000261662       
#> 2 eqtl-a-ENSG00000160460 ENSG00000160460       
#> 3 eqtl-a-ENSG00000205426 ENSG00000205426       
#> 4 ieu-a-9                Coronary heart disease
#> 5 eqtl-a-ENSG00000003436 ENSG00000003436       

# List the 6th - 10th Disease nodes, with full properties
meta_nodes_list_node(meta_node = "Disease", full_data = TRUE, limit = 5, offset = 0)

#> # A tibble: 5 x 9
#>   n.doid  n.umls  n.definition      n.id    n.label  n.icd9 n.mesh n.efo n.icd10
#>   <list>  <list>  <chr>             <chr>   <chr>    <list> <list> <lis> <list> 
#> 1 <chr [… <lgl [… A primary or met… http:/… cervica… <chr … <lgl … <lgl… <chr […
#> 2 <chr [… <chr [… A hordeolum that… http:/… hordeol… <chr … <lgl … <lgl… <chr […
#> 3 <chr [… <lgl [… A disease caused… http:/… Alkhurm… <lgl … <lgl … <lgl… <lgl […
#> 4 <chr [… <lgl [… Preeclampsia wit… http:/… severe … <chr … <lgl … <lgl… <lgl […
#> 5 <chr [… <chr [… A melanoma that … http:/… maligna… <lgl … <lgl … <lgl… <lgl […