List relationships under a meta relationship

GET /meta/rels/{meta_rel}/list

meta_rels_list_rel(meta_rel, limit = 10, offset = 0, mode = c("table", "raw"))

Arguments

meta_rel
Name of a meta relationship (e.g. MR). Use meta_rels_list to get the full list of meta relationships.

limit
Max number of items to retrieve.

offset
Number of items to skip. Use limit and offset in combination to do pagination.

mode
If mode = "table", returns a data frame (a tibble as per tidyverse convention). If mode = "raw", returns a raw response from EpiGraphDB API with minimal parsing done by httr.

Value

Data from GET /meta/rels/{meta_rel}/list

Examples

# List the first 5 MR relationships
meta_rels_list_rel(meta_rel = "MR_EVE_MR", limit = 5)

#> # A tibble: 5 x 42
#>   n.ncase  n._name n.year n.mr  n.author n.sex n.population n.sample_size n.nsnp
#>   <chr>    <chr>   <chr>  <chr> <chr>    <chr> <chr>        <chr>         <chr> 
#> 1 77099.0  Diseas… 2018.0 1     Neale l… Male… European     361194.0      13586…
#> 2 115893.0 Diseas… 2018.0 1     Neale l… Male… European     361194.0      13586…
#> 3 1281.0   Other/… 2018.0 1     Neale l… Male… European     361194.0      97040…
#> 4 1903.0   Ulcera… 2018.0 1     Neale l… Male… European     361194.0      10366…
#> 5 7233.0   Other … 2018.0 1     Neale l… Male… European     361194.0      12694…
#> # … with 33 more variables: n.build <chr>, n.ncontrol <chr>, n.trait <chr>,
#> #   n._source <list>, n.id <chr>, n._id <chr>, n.category <chr>, r.b <dbl>,
#> #   r.se <dbl>, r.method <chr>, r.nsnp <int>, r.selection <chr>, r.pval <dbl>,
#> #   r.ci_upp <dbl>, r._source <list>, r.ci_low <dbl>, r.moescore <dbl>,
#> #   m._name <chr>, m.year <chr>, m.mr <chr>, m.author <chr>, m.sex <chr>,
#> #   m.pmid <chr>, m.population <chr>, m.sample_size <chr>, m.nsnp <chr>,
#> #   m.build <chr>, m.trait <chr>, m._source <list>, m.id <chr>, m._id <chr>,
#> #   m.subcategory <chr>, m.category <chr>