Create an API request for the pQTL-related analyses

Create an API request for the pQTL-related analyses

pqtl_requests(query, rtype, pvalue, searchflag)

Arguments

query
(Required) A protein coding gene name or a trait name, eg. "ADAM19" or "Inflammatory bowel disease" which cannot be NULL.

rtype
(Optional) A type of data to be extracted, which can be one of these options:

simple: Basic summary

mrres: MR results (DEFAULT)

sglmr: Single SNP MR results

inst: SNP information

sense: Sensitivity analysis NOTE: mrres is a DEFAULT option.

pvalue
(Optional) A pvalue threshold for MR results with the DEFAULT set to 0.05. NOTE: this threshold applies to any rtype chosen.

searchflag
(Required) A flag to indicate whether you are searching for proteins or traits which cannot be NULL. If query is a protein name, then this flag should be "proteins"; if query is a trait, this flag should be "traits". NOTE: if the wrong flag is chosen for query, there will be no result returned.