MR evidence on confounding traits between exposure and outcome

MR evidence on confounding traits between exposure and outcome

confounder(exposure = NULL, outcome = NULL, type = c("confounder",
  "intermediate", "reverse_intermediate", "collider"),
  pval_threshold = 1e-05, mode = c("table", "raw"))

Arguments

exposure
A trait name, eg. "Body mass index", leaving exposure as NULL will return MR information related to a specific outcome. NOTE: exposure and outcome cannot be both NULL.

outcome
A trait name, eg. "Coronary heart disease", leaving outcome as NULL will return MR information related to a specific exposure. NOTE: exposure and outcome cannot be both NULL.

type
One in ["confounder", "intermediate", "reverse_intermediate", "collider"] Refer to the confounder view in web application for details

pval_threshold
P-value threshold

mode
If mode = "table", returns a data frame (a tibble as per tidyverse convention). If mode = "raw", returns a raw response from EpiGraphDB API with minimal parsing done by httr.

Value

Data from /confounder

Examples

confounder(exposure = "Body mass index", outcome = "Coronary heart disease")

#> # A tibble: 105 x 24
#>    exposure.id exposure.trait outcome.id outcome.trait cf.id cf.trait
#>    <chr>       <chr>          <chr>      <chr>         <chr> <chr>   
#>  1 835         Body mass ind… 7          Coronary hea… 60    Waist c…
#>  2 974         Body mass ind… 8          Coronary hea… UKB-… Treatme…
#>  3 95          Body mass ind… 7          Coronary hea… 51    Hip cir…
#>  4 95          Body mass ind… 7          Coronary hea… 63    Waist c…
#>  5 835         Body mass ind… 7          Coronary hea… 638   Asparag…
#>  6 95          Body mass ind… 7          Coronary hea… 62    Waist c…
#>  7 974         Body mass ind… 7          Coronary hea… 307   Cholest…
#>  8 974         Body mass ind… 8          Coronary hea… 307   Cholest…
#>  9 974         Body mass ind… 7          Coronary hea… 51    Hip cir…
#> 10 974         Body mass ind… 7          Coronary hea… 63    Waist c…
#> # … with 95 more rows, and 18 more variables: r1.estimate <dbl>, r1.se <dbl>,
#> #   r1.p <dbl>, r1.method <chr>, r1.selection <chr>, r1.moescore <dbl>,
#> #   r2.estimate <dbl>, r2.se <dbl>, r2.p <dbl>, r2.method <chr>,
#> #   r2.selection <chr>, r2.moescore <dbl>, r3.estimate <dbl>, r3.se <dbl>,
#> #   r3.p <dbl>, r3.method <chr>, r3.selection <chr>, r3.moescore <dbl>