List relationships under a meta relationship

GET /meta/rels/{meta_rel}/list

meta_rels_list_rel(meta_rel, limit = 10, offset = 0, mode = c("table", "raw"))

Arguments

meta_rel
Name of a meta relationship (e.g. MR). Use meta_rels_list to get the full list of meta relationships.

limit
Max number of items to retrieve.

offset
Number of items to skip. Use limit and offset in combination to do pagination.

mode
If mode = "table", returns a data frame (a tibble as per tidyverse convention). If mode = "raw", returns a raw response from EpiGraphDB API with minimal parsing done by httr.

Value

Data from GET /meta/rels/{meta_rel}/list

Examples

# List the first 5 MR relationships
meta_rels_list_rel(meta_rel = "MR", limit = 5)

#> # A tibble: 5 x 46
#>   n.note n.ncase n.access n.year n.mr  n.author n.consortium n.sex    n.priority
#>   <lgl>  <chr>   <chr>    <chr>  <chr> <chr>    <chr>        <chr>    <chr>     
#> 1 NA     8718    public   2017   1     Neale    Neale Lab    Males a… 1         
#> 2 NA     118     public   2017   1     Neale    Neale Lab    Males a… 1         
#> 3 NA     1489    public   2017   1     Neale    Neale Lab    Males a… 1         
#> 4 NA     2469    public   2017   1     Neale    Neale Lab    Males a… 1         
#> 5 NA     4609    public   2017   1     Neale    Neale Lab    Males a… 1         
#> # … with 37 more variables: n.population <chr>, n.unit <chr>,
#> #   n.sample_size <chr>, n.nsnp <chr>, n.ncontrol <chr>, n.trait <chr>,
#> #   n.id <chr>, n.subcategory <lgl>, n.category <lgl>, r.b <dbl>, r.se <dbl>,
#> #   r.method <chr>, r.nsnp <int>, r.selection <chr>, r.pval <dbl>,
#> #   r.ci_upp <dbl>, r.log10pval <dbl>, r.ci_low <dbl>, r.moescore <dbl>,
#> #   m.ncase <chr>, m.access <chr>, m.year <chr>, m.mr <chr>, m.author <chr>,
#> #   m.consortium <lgl>, m.sex <chr>, m.priority <chr>, m.pmid <chr>,
#> #   m.population <chr>, m.unit <lgl>, m.sample_size <chr>, m.nsnp <chr>,
#> #   m.ncontrol <chr>, m.trait <chr>, m.id <chr>, m.subcategory <chr>,
#> #   m.category <chr>