Return information related to the pQTL analysis

Return information related to the pQTL analysis

pqtl(query = NULL, rtype = "mrres", pvalue = 0.05,
  searchflag = NULL, mode = c("table", "raw"))

Arguments

query
(Required) A protein coding gene name or a trait name, eg. "ADAM19" or "Inflammatory bowel disease" which cannot be NULL.

rtype
(Optional) A type of data to be extracted, which can be one of these options:

simple: Basic summary

mrres: MR results (DEFAULT)

sglmr: Single SNP MR results

inst: SNP information

sense: Sensitivity analysis NOTE: mrres is a DEFAULT option.

pvalue
(Optional) A pvalue threshold for MR results with the DEFAULT set to 0.05. NOTE: this threshold applies to any rtype chosen.

searchflag
(Required) A flag to indicate whether you are searching for proteins or traits which cannot be NULL. If query is a protein name, then this flag should be "proteins"; if query is a trait, this flag should be "traits". NOTE: if the wrong flag is chosen for query, there will be no result returned.

mode
(Optional) If mode = "table", returns a data frame (a tibble as per tidyverse convention). If mode = "raw", returns a raw response from EpiGraphDB API with minimal parsing done by httr.

Examples

# Returns a data frame of MR results, while searching for proteins
pqtl(query="ADAM19", searchflag="proteins")

#> # A tibble: 16 x 8
#>    expID  outID             outID_mrbase  nsnp method      beta      se   pvalue
#>    <chr>  <chr>             <chr>        <int> <chr>      <dbl>   <dbl>    <dbl>
#>  1 ADAM19 Forced expirator… UKB-a:337        1 Wald r…  6.82e-2 8.92e-3 2.10e-14
#>  2 ADAM19 Height            89               1 Wald r…  4.88e-2 1.26e-2 1.08e- 4
#>  3 ADAM19 Non-cancer illne… UKB-a:66         1 Wald r… -1.26e-1 3.28e-2 1.21e- 4
#>  4 ADAM19 Hippocampus volu… 1045             1 Wald r…  6.37e+1 2.00e+1 1.48e- 3
#>  5 ADAM19 Forced vital cap… UKB-a:336        1 Wald r…  2.45e-2 8.46e-3 3.71e- 3
#>  6 ADAM19 Diagnoses - main… UKB-a:582        1 Wald r… -1.57e-1 5.85e-2 7.14e- 3
#>  7 ADAM19 Intracranial vol… 1041             1 Wald r…  1.99e+4 8.17e+3 1.50e- 2
#>  8 ADAM19 Neo-conscientiou… 114              1 Wald r… -7.48e-1 3.12e-1 1.66e- 2
#>  9 ADAM19 Subjective well … 1009             1 Wald r…  2.85e-2 1.22e-2 1.96e- 2
#> 10 ADAM19 Invasive mucinou… 1123             1 Wald r…  3.96e-1 1.73e-1 2.18e- 2
#> 11 ADAM19 HDL cholesterol   299              1 Wald r… -4.19e-2 1.99e-2 3.55e- 2
#> 12 ADAM19 Paget's disease   975              1 Wald r… -5.24e-1 2.49e-1 3.56e- 2
#> 13 ADAM19 Fractured bone s… UKB-a:441        1 Wald r… -1.04e-1 4.99e-2 3.69e- 2
#> 14 ADAM19 Diagnoses - main… UKB-a:559        1 Wald r… -1.77e-1 8.64e-2 4.09e- 2
#> 15 ADAM19 Systolic blood p… UKB-a:360        1 Wald r… -2.11e-2 1.05e-2 4.51e- 2
#> 16 ADAM19 Diagnoses - main… UKB-a:574        1 Wald r… -3.70e-1 1.85e-1 4.58e- 2

# Returns a data frame with SNP information, while searching for traits
pqtl(query="Inflammatory bowel disease", rtype="inst", searchflag="traits")

#> # A tibble: 56 x 13
#>    rsID  expID outID outID_mrbase ea    nea   eaf_exp sample_exp sample_out
#>    <chr> <chr> <chr> <chr>        <chr> <chr>   <dbl>      <dbl>      <dbl>
#>  1 rs11… IL23R Infl… 294          A     G      0.0669       3301      65642
#>  2 rs18… FCGR… Infl… 294          G     A      0.533        3301      65642
#>  3 rs49… IL12B Infl… 294          C     T      0.678        3301      65642
#>  4 rs11… C1QT… Infl… 294          A     G      0.472        3301      65642
#>  5 rs98… TXND… Infl… 294          T     G      0.110        3301      65642
#>  6 rs18… IL27  Infl… 294          T     G      0.327        3301      65642
#>  7 rs10… IL1R… Infl… 294          G     T      0.521        3301      65642
#>  8 rs12… FST   Infl… 294          T     C      0.44         3394      65642
#>  9 rs46… IGFB… Infl… 294          C     T      0.555        6861      65642
#> 10 rs12… KLKB1 Infl… 294          T     C      0.448        6861      65642
#> # … with 46 more rows, and 4 more variables: author_exp <chr>,
#> #   author_out <chr>, trans_cis <chr>, pvalue <dbl>

# Change a pvalue threshold (the default is 0.05)
pqtl(query="Inflammatory bowel disease", rtype="inst", pvalue=1.0, searchflag="traits")

#> # A tibble: 156 x 13
#>    rsID  expID outID outID_mrbase ea    nea   eaf_exp sample_exp sample_out
#>    <chr> <chr> <chr> <chr>        <chr> <chr>   <dbl>      <dbl>      <dbl>
#>  1 rs11… IL23R Infl… 294          A     G      0.0669       3301      65642
#>  2 rs18… FCGR… Infl… 294          G     A      0.533        3301      65642
#>  3 rs49… IL12B Infl… 294          C     T      0.678        3301      65642
#>  4 rs11… C1QT… Infl… 294          A     G      0.472        3301      65642
#>  5 rs98… TXND… Infl… 294          T     G      0.110        3301      65642
#>  6 rs18… IL27  Infl… 294          T     G      0.327        3301      65642
#>  7 rs10… IL1R… Infl… 294          G     T      0.521        3301      65642
#>  8 rs12… FST   Infl… 294          T     C      0.44         3394      65642
#>  9 rs46… IGFB… Infl… 294          C     T      0.555        6861      65642
#> 10 rs12… KLKB1 Infl… 294          T     C      0.448        6861      65642
#> # … with 146 more rows, and 4 more variables: author_exp <chr>,
#> #   author_out <chr>, trans_cis <chr>, pvalue <dbl>

# Returns raw response if mode="raw"
pqtl(
  query="ADAM19", searchflag="proteins",
  mode = "raw"
) %>% str()

#> List of 16
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Forced expiratory volume in 1-second (FEV1)"
#>   ..$ outID_mrbase: chr "UKB-a:337"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 0.0682
#>   ..$ se          : num 0.00892
#>   ..$ pvalue      : num 2.1e-14
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Height"
#>   ..$ outID_mrbase: chr "89"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 0.0488
#>   ..$ se          : num 0.0126
#>   ..$ pvalue      : num 0.000108
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Non-cancer illness code  self-reported: asthma"
#>   ..$ outID_mrbase: chr "UKB-a:66"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.126
#>   ..$ se          : num 0.0328
#>   ..$ pvalue      : num 0.000121
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Hippocampus volume"
#>   ..$ outID_mrbase: chr "1045"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 63.7
#>   ..$ se          : num 20
#>   ..$ pvalue      : num 0.00148
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Forced vital capacity (FVC)"
#>   ..$ outID_mrbase: chr "UKB-a:336"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 0.0245
#>   ..$ se          : num 0.00846
#>   ..$ pvalue      : num 0.00371
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Diagnoses - main ICD10: R10 Abdominal and pelvic pain"
#>   ..$ outID_mrbase: chr "UKB-a:582"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.157
#>   ..$ se          : num 0.0585
#>   ..$ pvalue      : num 0.00714
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Intracranial volume"
#>   ..$ outID_mrbase: chr "1041"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 19878
#>   ..$ se          : num 8174
#>   ..$ pvalue      : num 0.015
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Neo-conscientiousness"
#>   ..$ outID_mrbase: chr "114"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.748
#>   ..$ se          : num 0.312
#>   ..$ pvalue      : num 0.0166
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Subjective well being"
#>   ..$ outID_mrbase: chr "1009"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 0.0285
#>   ..$ se          : num 0.0122
#>   ..$ pvalue      : num 0.0196
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Invasive mucinous ovarian cancer"
#>   ..$ outID_mrbase: chr "1123"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num 0.396
#>   ..$ se          : num 0.173
#>   ..$ pvalue      : num 0.0218
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "HDL cholesterol"
#>   ..$ outID_mrbase: chr "299"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.0419
#>   ..$ se          : num 0.0199
#>   ..$ pvalue      : num 0.0355
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Paget's disease"
#>   ..$ outID_mrbase: chr "975"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.524
#>   ..$ se          : num 0.249
#>   ..$ pvalue      : num 0.0356
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Fractured bone site(s): Other bones"
#>   ..$ outID_mrbase: chr "UKB-a:441"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.104
#>   ..$ se          : num 0.0499
#>   ..$ pvalue      : num 0.0369
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Diagnoses - main ICD10: K80 Cholelithiasis"
#>   ..$ outID_mrbase: chr "UKB-a:559"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.177
#>   ..$ se          : num 0.0864
#>   ..$ pvalue      : num 0.0409
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Systolic blood pressure  automated reading"
#>   ..$ outID_mrbase: chr "UKB-a:360"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.0211
#>   ..$ se          : num 0.0105
#>   ..$ pvalue      : num 0.0451
#>  $ :List of 8
#>   ..$ expID       : chr "ADAM19"
#>   ..$ outID       : chr "Diagnoses - main ICD10: N20 Calculus of kidney and ureter"
#>   ..$ outID_mrbase: chr "UKB-a:574"
#>   ..$ nsnp        : int 1
#>   ..$ method      : chr "Wald ratio"
#>   ..$ beta        : num -0.37
#>   ..$ se          : num 0.185
#>   ..$ pvalue      : num 0.0458