Send data request to an EpiGraphDB API endpoint

This is a general purpose function to send data request which can be used when there has not been an R equivalent package function to an API endpoint. Underneath this is a wrapper around httr functions with better handling of returned status.

query_epigraphdb(
  route,
  params = NULL,
  mode = c("raw", "table"),
  method = c("GET", "POST"),
  retry_times = 3,
  retry_pause_min = 1
)

Arguments

route

An EpiGraphDB API endpoint route, e.g. "/mr" or "/confounder". Consult the EpiGraphDB API documentation.

params

A list of parameters associated with the query endpoint.

mode

c("raw", "table"), if "table" then the query handler will try to convert the returned data to a tibble dataframe. NOTE: The default mode is "raw" which will NOT convert the returned response to a dataframe. This is different to functions that query topic endpoints which default to return a dataframe. Explicitly specify mode = "table" when needed.

method

Type of HTTP (GET, POST, PUT, etc.) method.

NOTE: When sending a POST request where a specific parameter is specified as a list on the API, and if the equivalent in R is a vector of length 1, you should wrap this parameter in I(), e.g. I(c("APOE")) to avoid auto unboxing. For details, please refer to httr::POST

retry_times

Number of times the function will retry the request to the API.

retry_pause_min

Minimum number of seconds to wait for the next retry.

Value

Data from an EpiGraphDB API endpoint.

Examples

# GET /mr
# equivalent to `mr(exposure_trait = "Body mass index", outcome_trait = "Coronary heart disease")`
# \dontrun{
query_epigraphdb(
  route = "/mr",
  params = list(
    exposure_trait = "Body mass index",
    outcome_trait = "Coronary heart disease"
  ),
  mode = "table"
)
#> # A tibble: 18 × 10
#>    exposure.id  exposure.trait  outcome.id  outcome.trait   mr.b  mr.se  mr.pval
#>    <chr>        <chr>           <chr>       <chr>          <dbl>  <dbl>    <dbl>
#>  1 ieu-a-2      Body mass index ieu-a-7     Coronary hear… 0.464 0.0415 5.46e-29
#>  2 ebi-a-GCST0… Body mass index ieu-a-7     Coronary hear… 0.457 0.0410 3.33e-20
#>  3 ieu-a-974    Body mass index ieu-a-7     Coronary hear… 0.389 0.0493 3.42e-15
#>  4 ieu-a-835    Body mass index ieu-a-7     Coronary hear… 0.417 0.0492 1.00e-11
#>  5 ieu-a-974    Body mass index ieu-a-9     Coronary hear… 0.320 0.0536 2.32e- 9
#>  6 ieu-a-2      Body mass index ieu-a-9     Coronary hear… 0.358 0.0535 5.91e- 9
#>  7 ieu-a-835    Body mass index ieu-a-9     Coronary hear… 0.397 0.0604 1.79e- 8
#>  8 ebi-a-GCST0… Body mass index ieu-a-9     Coronary hear… 0.341 0.0590 1.24e- 7
#>  9 ieu-a-95     Body mass index ieu-a-9     Coronary hear… 0.371 0.0708 1.62e- 7
#> 10 ebi-a-GCST0… Body mass index ieu-a-6     Coronary hear… 0.493 0.0986 5.88e- 7
#> 11 ieu-a-785    Body mass index ieu-a-9     Coronary hear… 0.395 0.0609 1.07e- 6
#> 12 ebi-a-GCST0… Body mass index ebi-a-GCST… Coronary hear… 0.309 0.0648 1.81e- 6
#> 13 ebi-a-GCST0… Body mass index ieu-a-8     Coronary hear… 0.309 0.0648 1.81e- 6
#> 14 ebi-a-GCST0… Body mass index ieu-a-7     Coronary hear… 0.275 0.0514 2.76e- 6
#> 15 ieu-a-95     Body mass index ieu-a-7     Coronary hear… 0.455 0.0971 2.82e- 6
#> 16 ieu-a-2      Body mass index ieu-a-8     Coronary hear… 0.317 0.0686 3.93e- 6
#> 17 ieu-a-2      Body mass index ebi-a-GCST… Coronary hear… 0.312 0.0688 5.86e- 6
#> 18 ieu-a-974    Body mass index ieu-a-8     Coronary hear… 0.328 0.0731 6.97e- 6
#> # … with 3 more variables: mr.method <chr>, mr.selection <chr>,
#> #   mr.moescore <dbl>
# }

# GET /meta/nodes/Gwas/list
# \dontrun{
query_epigraphdb(
  route = "/meta/nodes/Gwas/list",
  params = list(
    limit = 5,
    offset = 0
  )
) %>% str(1)
#> List of 2
#>  $ metadata:List of 3
#>  $ results :List of 5
# }

# POST /protein/ppi
# \dontrun{
query_epigraphdb(
  route = "/protein/ppi",
  params = list(
    uniprot_id_list = c("P30793", "Q9NZM1", "O95236")
  ),
  method = "POST"
)
#> $metadata
#> $metadata$query
#> [1] "MATCH p=(protein:Protein)-[r:STRING_INTERACT_WITH]-(assoc_protein:Protein) WHERE protein.uniprot_id IN ['P30793', 'Q9NZM1', 'O95236'] RETURN protein {.uniprot_id}, assoc_protein {.uniprot_id}"
#> 
#> $metadata$total_seconds
#> [1] 0.013172
#> 
#> $metadata$empty_results
#> [1] FALSE
#> 
#> 
#> $results
#> $results[[1]]
#> $results[[1]]$protein
#> $results[[1]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[1]]$assoc_protein
#> $results[[1]]$assoc_protein$uniprot_id
#> [1] "O14638"
#> 
#> 
#> 
#> $results[[2]]
#> $results[[2]]$protein
#> $results[[2]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[2]]$assoc_protein
#> $results[[2]]$assoc_protein$uniprot_id
#> [1] "P20711"
#> 
#> 
#> 
#> $results[[3]]
#> $results[[3]]$protein
#> $results[[3]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[3]]$assoc_protein
#> $results[[3]]$assoc_protein$uniprot_id
#> [1] "O43556"
#> 
#> 
#> 
#> $results[[4]]
#> $results[[4]]$protein
#> $results[[4]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[4]]$assoc_protein
#> $results[[4]]$assoc_protein$uniprot_id
#> [1] "P00374"
#> 
#> 
#> 
#> $results[[5]]
#> $results[[5]]$protein
#> $results[[5]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[5]]$assoc_protein
#> $results[[5]]$assoc_protein$uniprot_id
#> [1] "Q32Q12"
#> 
#> 
#> 
#> $results[[6]]
#> $results[[6]]$protein
#> $results[[6]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[6]]$assoc_protein
#> $results[[6]]$assoc_protein$uniprot_id
#> [1] "P22392"
#> 
#> 
#> 
#> $results[[7]]
#> $results[[7]]$protein
#> $results[[7]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[7]]$assoc_protein
#> $results[[7]]$assoc_protein$uniprot_id
#> [1] "P05187"
#> 
#> 
#> 
#> $results[[8]]
#> $results[[8]]$protein
#> $results[[8]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[8]]$assoc_protein
#> $results[[8]]$assoc_protein$uniprot_id
#> [1] "P22102"
#> 
#> 
#> 
#> $results[[9]]
#> $results[[9]]$protein
#> $results[[9]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[9]]$assoc_protein
#> $results[[9]]$assoc_protein$uniprot_id
#> [1] "P07101"
#> 
#> 
#> 
#> $results[[10]]
#> $results[[10]]$protein
#> $results[[10]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[10]]$assoc_protein
#> $results[[10]]$assoc_protein$uniprot_id
#> [1] "Q9BY32"
#> 
#> 
#> 
#> $results[[11]]
#> $results[[11]]$protein
#> $results[[11]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[11]]$assoc_protein
#> $results[[11]]$assoc_protein$uniprot_id
#> [1] "P50583"
#> 
#> 
#> 
#> $results[[12]]
#> $results[[12]]$protein
#> $results[[12]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[12]]$assoc_protein
#> $results[[12]]$assoc_protein$uniprot_id
#> [1] "O95197"
#> 
#> 
#> 
#> $results[[13]]
#> $results[[13]]$protein
#> $results[[13]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[13]]$assoc_protein
#> $results[[13]]$assoc_protein$uniprot_id
#> [1] "P05186"
#> 
#> 
#> 
#> $results[[14]]
#> $results[[14]]$protein
#> $results[[14]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[14]]$assoc_protein
#> $results[[14]]$assoc_protein$uniprot_id
#> [1] "Q05932"
#> 
#> 
#> 
#> $results[[15]]
#> $results[[15]]$protein
#> $results[[15]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[15]]$assoc_protein
#> $results[[15]]$assoc_protein$uniprot_id
#> [1] "Q9NZB8"
#> 
#> 
#> 
#> $results[[16]]
#> $results[[16]]$protein
#> $results[[16]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[16]]$assoc_protein
#> $results[[16]]$assoc_protein$uniprot_id
#> [1] "Q5MY95"
#> 
#> 
#> 
#> $results[[17]]
#> $results[[17]]$protein
#> $results[[17]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[17]]$assoc_protein
#> $results[[17]]$assoc_protein$uniprot_id
#> [1] "Q9BX66"
#> 
#> 
#> 
#> $results[[18]]
#> $results[[18]]$protein
#> $results[[18]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[18]]$assoc_protein
#> $results[[18]]$assoc_protein$uniprot_id
#> [1] "P49961"
#> 
#> 
#> 
#> $results[[19]]
#> $results[[19]]$protein
#> $results[[19]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[19]]$assoc_protein
#> $results[[19]]$assoc_protein$uniprot_id
#> [1] "P16066"
#> 
#> 
#> 
#> $results[[20]]
#> $results[[20]]$protein
#> $results[[20]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[20]]$assoc_protein
#> $results[[20]]$assoc_protein$uniprot_id
#> [1] "Q9Y5B8"
#> 
#> 
#> 
#> $results[[21]]
#> $results[[21]]$protein
#> $results[[21]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[21]]$assoc_protein
#> $results[[21]]$assoc_protein$uniprot_id
#> [1] "Q9NP79"
#> 
#> 
#> 
#> $results[[22]]
#> $results[[22]]$protein
#> $results[[22]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[22]]$assoc_protein
#> $results[[22]]$assoc_protein$uniprot_id
#> [1] "P22413"
#> 
#> 
#> 
#> $results[[23]]
#> $results[[23]]$protein
#> $results[[23]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[23]]$assoc_protein
#> $results[[23]]$assoc_protein$uniprot_id
#> [1] "O94875"
#> 
#> 
#> 
#> $results[[24]]
#> $results[[24]]$protein
#> $results[[24]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[24]]$assoc_protein
#> $results[[24]]$assoc_protein$uniprot_id
#> [1] "O14656"
#> 
#> 
#> 
#> $results[[25]]
#> $results[[25]]$protein
#> $results[[25]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[25]]$assoc_protein
#> $results[[25]]$assoc_protein$uniprot_id
#> [1] "P20594"
#> 
#> 
#> 
#> $results[[26]]
#> $results[[26]]$protein
#> $results[[26]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[26]]$assoc_protein
#> $results[[26]]$assoc_protein$uniprot_id
#> [1] "P30613"
#> 
#> 
#> 
#> $results[[27]]
#> $results[[27]]$protein
#> $results[[27]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[27]]$assoc_protein
#> $results[[27]]$assoc_protein$uniprot_id
#> [1] "Q9NQC3"
#> 
#> 
#> 
#> $results[[28]]
#> $results[[28]]$protein
#> $results[[28]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[28]]$assoc_protein
#> $results[[28]]$assoc_protein$uniprot_id
#> [1] "P15531"
#> 
#> 
#> 
#> $results[[29]]
#> $results[[29]]$protein
#> $results[[29]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[29]]$assoc_protein
#> $results[[29]]$assoc_protein$uniprot_id
#> [1] "P14618"
#> 
#> 
#> 
#> $results[[30]]
#> $results[[30]]$protein
#> $results[[30]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[30]]$assoc_protein
#> $results[[30]]$assoc_protein$uniprot_id
#> [1] "O75355"
#> 
#> 
#> 
#> $results[[31]]
#> $results[[31]]$protein
#> $results[[31]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[31]]$assoc_protein
#> $results[[31]]$assoc_protein$uniprot_id
#> [1] "Q02108"
#> 
#> 
#> 
#> $results[[32]]
#> $results[[32]]$protein
#> $results[[32]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[32]]$assoc_protein
#> $results[[32]]$assoc_protein$uniprot_id
#> [1] "P09923"
#> 
#> 
#> 
#> $results[[33]]
#> $results[[33]]$protein
#> $results[[33]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[33]]$assoc_protein
#> $results[[33]]$assoc_protein$uniprot_id
#> [1] "P10696"
#> 
#> 
#> 
#> $results[[34]]
#> $results[[34]]$protein
#> $results[[34]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[34]]$assoc_protein
#> $results[[34]]$assoc_protein$uniprot_id
#> [1] "P16152"
#> 
#> 
#> 
#> $results[[35]]
#> $results[[35]]$protein
#> $results[[35]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[35]]$assoc_protein
#> $results[[35]]$assoc_protein$uniprot_id
#> [1] "P33402"
#> 
#> 
#> 
#> $results[[36]]
#> $results[[36]]$protein
#> $results[[36]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[36]]$assoc_protein
#> $results[[36]]$assoc_protein$uniprot_id
#> [1] "P09417"
#> 
#> 
#> 
#> $results[[37]]
#> $results[[37]]$protein
#> $results[[37]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[37]]$assoc_protein
#> $results[[37]]$assoc_protein$uniprot_id
#> [1] "Q03393"
#> 
#> 
#> 
#> $results[[38]]
#> $results[[38]]$protein
#> $results[[38]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[38]]$assoc_protein
#> $results[[38]]$assoc_protein$uniprot_id
#> [1] "Q16799"
#> 
#> 
#> 
#> $results[[39]]
#> $results[[39]]$protein
#> $results[[39]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[39]]$assoc_protein
#> $results[[39]]$assoc_protein$uniprot_id
#> [1] "Q02153"
#> 
#> 
#> 
#> $results[[40]]
#> $results[[40]]$protein
#> $results[[40]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[40]]$assoc_protein
#> $results[[40]]$assoc_protein$uniprot_id
#> [1] "P25092"
#> 
#> 
#> 
#> $results[[41]]
#> $results[[41]]$protein
#> $results[[41]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[41]]$assoc_protein
#> $results[[41]]$assoc_protein$uniprot_id
#> [1] "P30047"
#> 
#> 
#> 
#> $results[[42]]
#> $results[[42]]$protein
#> $results[[42]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[42]]$assoc_protein
#> $results[[42]]$assoc_protein$uniprot_id
#> [1] "Q02846"
#> 
#> 
#> 
#> $results[[43]]
#> $results[[43]]$protein
#> $results[[43]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[43]]$assoc_protein
#> $results[[43]]$assoc_protein$uniprot_id
#> [1] "Q9NVV9"
#> 
#> 
#> 
#> $results[[44]]
#> $results[[44]]$protein
#> $results[[44]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[44]]$assoc_protein
#> $results[[44]]$assoc_protein$uniprot_id
#> [1] "P17752"
#> 
#> 
#> 
#> $results[[45]]
#> $results[[45]]$protein
#> $results[[45]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[45]]$assoc_protein
#> $results[[45]]$assoc_protein$uniprot_id
#> [1] "O75298"
#> 
#> 
#> 
#> $results[[46]]
#> $results[[46]]$protein
#> $results[[46]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[46]]$assoc_protein
#> $results[[46]]$assoc_protein$uniprot_id
#> [1] "Q96PE7"
#> 
#> 
#> 
#> $results[[47]]
#> $results[[47]]$protein
#> $results[[47]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[47]]$assoc_protein
#> $results[[47]]$assoc_protein$uniprot_id
#> [1] "P35270"
#> 
#> 
#> 
#> $results[[48]]
#> $results[[48]]$protein
#> $results[[48]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[48]]$assoc_protein
#> $results[[48]]$assoc_protein$uniprot_id
#> [1] "O00746"
#> 
#> 
#> 
#> $results[[49]]
#> $results[[49]]$protein
#> $results[[49]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[49]]$assoc_protein
#> $results[[49]]$assoc_protein$uniprot_id
#> [1] "Q13232"
#> 
#> 
#> 
#> $results[[50]]
#> $results[[50]]$protein
#> $results[[50]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[50]]$assoc_protein
#> $results[[50]]$assoc_protein$uniprot_id
#> [1] "P51841"
#> 
#> 
#> 
#> $results[[51]]
#> $results[[51]]$protein
#> $results[[51]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[51]]$assoc_protein
#> $results[[51]]$assoc_protein$uniprot_id
#> [1] "P00439"
#> 
#> 
#> 
#> $results[[52]]
#> $results[[52]]$protein
#> $results[[52]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[52]]$assoc_protein
#> $results[[52]]$assoc_protein$uniprot_id
#> [1] "O75414"
#> 
#> 
#> 
#> $results[[53]]
#> $results[[53]]$protein
#> $results[[53]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[53]]$assoc_protein
#> $results[[53]]$assoc_protein$uniprot_id
#> [1] "Q5TCS8"
#> 
#> 
#> 
#> $results[[54]]
#> $results[[54]]$protein
#> $results[[54]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[54]]$assoc_protein
#> $results[[54]]$assoc_protein$uniprot_id
#> [1] "Q02846"
#> 
#> 
#> 
#> $results[[55]]
#> $results[[55]]$protein
#> $results[[55]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[55]]$assoc_protein
#> $results[[55]]$assoc_protein$uniprot_id
#> [1] "P50583"
#> 
#> 
#> 
#> $results[[56]]
#> $results[[56]]$protein
#> $results[[56]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[56]]$assoc_protein
#> $results[[56]]$assoc_protein$uniprot_id
#> [1] "P35270"
#> 
#> 
#> 
#> $results[[57]]
#> $results[[57]]$protein
#> $results[[57]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[57]]$assoc_protein
#> $results[[57]]$assoc_protein$uniprot_id
#> [1] "Q9NP79"
#> 
#> 
#> 
#> $results[[58]]
#> $results[[58]]$protein
#> $results[[58]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[58]]$assoc_protein
#> $results[[58]]$assoc_protein$uniprot_id
#> [1] "O43556"
#> 
#> 
#> 
#> $results[[59]]
#> $results[[59]]$protein
#> $results[[59]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[59]]$assoc_protein
#> $results[[59]]$assoc_protein$uniprot_id
#> [1] "P30613"
#> 
#> 
#> 
#> $results[[60]]
#> $results[[60]]$protein
#> $results[[60]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[60]]$assoc_protein
#> $results[[60]]$assoc_protein$uniprot_id
#> [1] "P15531"
#> 
#> 
#> 
#> $results[[61]]
#> $results[[61]]$protein
#> $results[[61]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[61]]$assoc_protein
#> $results[[61]]$assoc_protein$uniprot_id
#> [1] "Q05932"
#> 
#> 
#> 
#> $results[[62]]
#> $results[[62]]$protein
#> $results[[62]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[62]]$assoc_protein
#> $results[[62]]$assoc_protein$uniprot_id
#> [1] "P07101"
#> 
#> 
#> 
#> $results[[63]]
#> $results[[63]]$protein
#> $results[[63]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[63]]$assoc_protein
#> $results[[63]]$assoc_protein$uniprot_id
#> [1] "Q16799"
#> 
#> 
#> 
#> $results[[64]]
#> $results[[64]]$protein
#> $results[[64]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[64]]$assoc_protein
#> $results[[64]]$assoc_protein$uniprot_id
#> [1] "P10696"
#> 
#> 
#> 
#> $results[[65]]
#> $results[[65]]$protein
#> $results[[65]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[65]]$assoc_protein
#> $results[[65]]$assoc_protein$uniprot_id
#> [1] "P05186"
#> 
#> 
#> 
#> $results[[66]]
#> $results[[66]]$protein
#> $results[[66]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[66]]$assoc_protein
#> $results[[66]]$assoc_protein$uniprot_id
#> [1] "P09923"
#> 
#> 
#> 
#> $results[[67]]
#> $results[[67]]$protein
#> $results[[67]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[67]]$assoc_protein
#> $results[[67]]$assoc_protein$uniprot_id
#> [1] "P49961"
#> 
#> 
#> 
#> $results[[68]]
#> $results[[68]]$protein
#> $results[[68]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[68]]$assoc_protein
#> $results[[68]]$assoc_protein$uniprot_id
#> [1] "O75355"
#> 
#> 
#> 
#> $results[[69]]
#> $results[[69]]$protein
#> $results[[69]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[69]]$assoc_protein
#> $results[[69]]$assoc_protein$uniprot_id
#> [1] "Q9BY32"
#> 
#> 
#> 
#> $results[[70]]
#> $results[[70]]$protein
#> $results[[70]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[70]]$assoc_protein
#> $results[[70]]$assoc_protein$uniprot_id
#> [1] "O75298"
#> 
#> 
#> 
#> $results[[71]]
#> $results[[71]]$protein
#> $results[[71]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[71]]$assoc_protein
#> $results[[71]]$assoc_protein$uniprot_id
#> [1] "P05187"
#> 
#> 
#> 
#> $results[[72]]
#> $results[[72]]$protein
#> $results[[72]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[72]]$assoc_protein
#> $results[[72]]$assoc_protein$uniprot_id
#> [1] "Q9NVV9"
#> 
#> 
#> 
#> $results[[73]]
#> $results[[73]]$protein
#> $results[[73]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[73]]$assoc_protein
#> $results[[73]]$assoc_protein$uniprot_id
#> [1] "P20711"
#> 
#> 
#> 
#> $results[[74]]
#> $results[[74]]$protein
#> $results[[74]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[74]]$assoc_protein
#> $results[[74]]$assoc_protein$uniprot_id
#> [1] "O94875"
#> 
#> 
#> 
#> $results[[75]]
#> $results[[75]]$protein
#> $results[[75]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[75]]$assoc_protein
#> $results[[75]]$assoc_protein$uniprot_id
#> [1] "Q9BX66"
#> 
#> 
#> 
#> $results[[76]]
#> $results[[76]]$protein
#> $results[[76]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[76]]$assoc_protein
#> $results[[76]]$assoc_protein$uniprot_id
#> [1] "P00374"
#> 
#> 
#> 
#> $results[[77]]
#> $results[[77]]$protein
#> $results[[77]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[77]]$assoc_protein
#> $results[[77]]$assoc_protein$uniprot_id
#> [1] "Q02153"
#> 
#> 
#> 
#> $results[[78]]
#> $results[[78]]$protein
#> $results[[78]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[78]]$assoc_protein
#> $results[[78]]$assoc_protein$uniprot_id
#> [1] "P20594"
#> 
#> 
#> 
#> $results[[79]]
#> $results[[79]]$protein
#> $results[[79]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[79]]$assoc_protein
#> $results[[79]]$assoc_protein$uniprot_id
#> [1] "Q96PE7"
#> 
#> 
#> 
#> $results[[80]]
#> $results[[80]]$protein
#> $results[[80]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[80]]$assoc_protein
#> $results[[80]]$assoc_protein$uniprot_id
#> [1] "P09417"
#> 
#> 
#> 
#> $results[[81]]
#> $results[[81]]$protein
#> $results[[81]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[81]]$assoc_protein
#> $results[[81]]$assoc_protein$uniprot_id
#> [1] "P17752"
#> 
#> 
#> 
#> $results[[82]]
#> $results[[82]]$protein
#> $results[[82]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[82]]$assoc_protein
#> $results[[82]]$assoc_protein$uniprot_id
#> [1] "Q03393"
#> 
#> 
#> 
#> $results[[83]]
#> $results[[83]]$protein
#> $results[[83]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[83]]$assoc_protein
#> $results[[83]]$assoc_protein$uniprot_id
#> [1] "O95197"
#> 
#> 
#> 
#> $results[[84]]
#> $results[[84]]$protein
#> $results[[84]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[84]]$assoc_protein
#> $results[[84]]$assoc_protein$uniprot_id
#> [1] "Q9Y5B8"
#> 
#> 
#> 
#> $results[[85]]
#> $results[[85]]$protein
#> $results[[85]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[85]]$assoc_protein
#> $results[[85]]$assoc_protein$uniprot_id
#> [1] "P22102"
#> 
#> 
#> 
#> $results[[86]]
#> $results[[86]]$protein
#> $results[[86]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[86]]$assoc_protein
#> $results[[86]]$assoc_protein$uniprot_id
#> [1] "O14638"
#> 
#> 
#> 
#> $results[[87]]
#> $results[[87]]$protein
#> $results[[87]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[87]]$assoc_protein
#> $results[[87]]$assoc_protein$uniprot_id
#> [1] "P22413"
#> 
#> 
#> 
#> $results[[88]]
#> $results[[88]]$protein
#> $results[[88]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[88]]$assoc_protein
#> $results[[88]]$assoc_protein$uniprot_id
#> [1] "P33402"
#> 
#> 
#> 
#> $results[[89]]
#> $results[[89]]$protein
#> $results[[89]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[89]]$assoc_protein
#> $results[[89]]$assoc_protein$uniprot_id
#> [1] "Q02108"
#> 
#> 
#> 
#> $results[[90]]
#> $results[[90]]$protein
#> $results[[90]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[90]]$assoc_protein
#> $results[[90]]$assoc_protein$uniprot_id
#> [1] "P16066"
#> 
#> 
#> 
#> $results[[91]]
#> $results[[91]]$protein
#> $results[[91]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[91]]$assoc_protein
#> $results[[91]]$assoc_protein$uniprot_id
#> [1] "Q5MY95"
#> 
#> 
#> 
#> $results[[92]]
#> $results[[92]]$protein
#> $results[[92]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[92]]$assoc_protein
#> $results[[92]]$assoc_protein$uniprot_id
#> [1] "P51841"
#> 
#> 
#> 
#> $results[[93]]
#> $results[[93]]$protein
#> $results[[93]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[93]]$assoc_protein
#> $results[[93]]$assoc_protein$uniprot_id
#> [1] "O14656"
#> 
#> 
#> 
#> $results[[94]]
#> $results[[94]]$protein
#> $results[[94]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[94]]$assoc_protein
#> $results[[94]]$assoc_protein$uniprot_id
#> [1] "P16152"
#> 
#> 
#> 
#> $results[[95]]
#> $results[[95]]$protein
#> $results[[95]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[95]]$assoc_protein
#> $results[[95]]$assoc_protein$uniprot_id
#> [1] "P30047"
#> 
#> 
#> 
#> $results[[96]]
#> $results[[96]]$protein
#> $results[[96]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[96]]$assoc_protein
#> $results[[96]]$assoc_protein$uniprot_id
#> [1] "Q13232"
#> 
#> 
#> 
#> $results[[97]]
#> $results[[97]]$protein
#> $results[[97]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[97]]$assoc_protein
#> $results[[97]]$assoc_protein$uniprot_id
#> [1] "P25092"
#> 
#> 
#> 
#> $results[[98]]
#> $results[[98]]$protein
#> $results[[98]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[98]]$assoc_protein
#> $results[[98]]$assoc_protein$uniprot_id
#> [1] "Q5TCS8"
#> 
#> 
#> 
#> $results[[99]]
#> $results[[99]]$protein
#> $results[[99]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[99]]$assoc_protein
#> $results[[99]]$assoc_protein$uniprot_id
#> [1] "O00746"
#> 
#> 
#> 
#> $results[[100]]
#> $results[[100]]$protein
#> $results[[100]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[100]]$assoc_protein
#> $results[[100]]$assoc_protein$uniprot_id
#> [1] "O75414"
#> 
#> 
#> 
#> $results[[101]]
#> $results[[101]]$protein
#> $results[[101]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[101]]$assoc_protein
#> $results[[101]]$assoc_protein$uniprot_id
#> [1] "P22392"
#> 
#> 
#> 
#> $results[[102]]
#> $results[[102]]$protein
#> $results[[102]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[102]]$assoc_protein
#> $results[[102]]$assoc_protein$uniprot_id
#> [1] "Q32Q12"
#> 
#> 
#> 
#> $results[[103]]
#> $results[[103]]$protein
#> $results[[103]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[103]]$assoc_protein
#> $results[[103]]$assoc_protein$uniprot_id
#> [1] "P00439"
#> 
#> 
#> 
#> $results[[104]]
#> $results[[104]]$protein
#> $results[[104]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[104]]$assoc_protein
#> $results[[104]]$assoc_protein$uniprot_id
#> [1] "P14618"
#> 
#> 
#> 
#> $results[[105]]
#> $results[[105]]$protein
#> $results[[105]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[105]]$assoc_protein
#> $results[[105]]$assoc_protein$uniprot_id
#> [1] "Q9NQC3"
#> 
#> 
#> 
#> $results[[106]]
#> $results[[106]]$protein
#> $results[[106]]$protein$uniprot_id
#> [1] "P30793"
#> 
#> 
#> $results[[106]]$assoc_protein
#> $results[[106]]$assoc_protein$uniprot_id
#> [1] "Q9NZB8"
#> 
#> 
#> 
#> $results[[107]]
#> $results[[107]]$protein
#> $results[[107]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[107]]$assoc_protein
#> $results[[107]]$assoc_protein$uniprot_id
#> [1] "Q9NZN3"
#> 
#> 
#> 
#> $results[[108]]
#> $results[[108]]$protein
#> $results[[108]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[108]]$assoc_protein
#> $results[[108]]$assoc_protein$uniprot_id
#> [1] "P15692"
#> 
#> 
#> 
#> $results[[109]]
#> $results[[109]]$protein
#> $results[[109]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[109]]$assoc_protein
#> $results[[109]]$assoc_protein$uniprot_id
#> [1] "P50570"
#> 
#> 
#> 
#> $results[[110]]
#> $results[[110]]$protein
#> $results[[110]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[110]]$assoc_protein
#> $results[[110]]$assoc_protein$uniprot_id
#> [1] "Q9NZN3"
#> 
#> 
#> 
#> $results[[111]]
#> $results[[111]]$protein
#> $results[[111]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[111]]$assoc_protein
#> $results[[111]]$assoc_protein$uniprot_id
#> [1] "P22681"
#> 
#> 
#> 
#> $results[[112]]
#> $results[[112]]$protein
#> $results[[112]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[112]]$assoc_protein
#> $results[[112]]$assoc_protein$uniprot_id
#> [1] "P35968"
#> 
#> 
#> 
#> $results[[113]]
#> $results[[113]]$protein
#> $results[[113]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[113]]$assoc_protein
#> $results[[113]]$assoc_protein$uniprot_id
#> [1] "Q9NZN4"
#> 
#> 
#> 
#> $results[[114]]
#> $results[[114]]$protein
#> $results[[114]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[114]]$assoc_protein
#> $results[[114]]$assoc_protein$uniprot_id
#> [1] "Q9NZN4"
#> 
#> 
#> 
#> $results[[115]]
#> $results[[115]]$protein
#> $results[[115]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[115]]$assoc_protein
#> $results[[115]]$assoc_protein$uniprot_id
#> [1] "P15692"
#> 
#> 
#> 
#> $results[[116]]
#> $results[[116]]$protein
#> $results[[116]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[116]]$assoc_protein
#> $results[[116]]$assoc_protein$uniprot_id
#> [1] "P35968"
#> 
#> 
#> 
#> $results[[117]]
#> $results[[117]]$protein
#> $results[[117]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[117]]$assoc_protein
#> $results[[117]]$assoc_protein$uniprot_id
#> [1] "P50570"
#> 
#> 
#> 
#> $results[[118]]
#> $results[[118]]$protein
#> $results[[118]]$protein$uniprot_id
#> [1] "Q9NZM1"
#> 
#> 
#> $results[[118]]$assoc_protein
#> $results[[118]]$assoc_protein$uniprot_id
#> [1] "P22681"
#> 
#> 
#> 
#> 
# }

# error handling
# \dontrun{
tryCatch(
  query_epigraphdb(
    route = "/mr",
    params = list(
      exposure_trait = NULL,
      outcome_trait = NULL
    ),
    retry_times = 0
  ),
  error = function(e) {
    message(e)
  }
)
#> Error in stop_for_status(response = response, context = list(params = params,     url = url)): HTTP error: Unprocessable Entity (WebDAV; RFC 4918) (status code 422).
#> Detail:
#> $detail
#> [1] "At least one of the following parameters should be non-missing: [exposure_trait, outcome_trait]"
#> 
#> Context:
#> $params
#> $params$exposure_trait
#> NULL
#> 
#> $params$outcome_trait
#> NULL
#> 
#> 
#> $url
#> https://api.epigraphdb.org/mr
# }